Reconcile the sgi-demos ports' copies into this fork.
- libtess internals back to GLdouble (standard GLU ABI; the public tess
  API already was) - float internals broke callers and tess precision
- GLUES_USE_HW_MIPMAP: POT gluBuild2DMipmaps via
  gl4es GL_GENERATE_MIPMAP; CPU chains sample black under gl4es
- GLUES_GL4ES: add glext.h; self-include glu_mangle.h on Mac/
  Emscripten so glu* definitions mangle without build-system flags
- README: this fork is the shared canonical GLU for the sgi-demos ports

+++ source/libtess/geom.c
@@ -45,7 +45,7 @@ int __gl_vertLeq(GLUvertex* u, GLUvertex* v)
    return VertLeq(u, v);
 }
 
-GLfloat __gl_edgeEval(GLUvertex* u, GLUvertex* v, GLUvertex* w)
+GLdouble __gl_edgeEval(GLUvertex* u, GLUvertex* v, GLUvertex* w)
 {
    /* Given three vertices u,v,w such that VertLeq(u,v) && VertLeq(v,w),
     * evaluates the t-coord of the edge uw at the s-coord of the vertex v.
@@ -57,7 +57,7 @@ GLfloat __gl_edgeEval(GLUvertex* u, GLUvertex* v, GLUvertex* w)
     * let r be the negated result (this evaluates (uw)(v->s)), then
     * r is guaranteed to satisfy MIN(u->t,w->t) <= r <= MAX(u->t,w->t).
     */
-   GLfloat gapL, gapR;
+   GLdouble gapL, gapR;
 
    assert(VertLeq(u, v) && VertLeq(v, w));
 
@@ -80,13 +80,13 @@ GLfloat __gl_edgeEval(GLUvertex* u, GLUvertex* v, GLUvertex* w)
    return 0;
 }
 
-GLfloat __gl_edgeSign(GLUvertex* u, GLUvertex* v, GLUvertex* w)
+GLdouble __gl_edgeSign(GLUvertex* u, GLUvertex* v, GLUvertex* w)
 {
    /* Returns a number whose sign matches EdgeEval(u,v,w) but which
     * is cheaper to evaluate.  Returns > 0, == 0 , or < 0
     * as v is above, on, or below the edge uw.
     */
-   GLfloat gapL, gapR;
+   GLdouble gapL, gapR;
 
    assert(VertLeq(u, v) && VertLeq(v, w));
 
@@ -106,7 +106,7 @@ GLfloat __gl_edgeSign(GLUvertex* u, GLUvertex* v, GLUvertex* w)
  * Define versions of EdgeSign, EdgeEval with s and t transposed.
  */
 
-GLfloat __gl_transEval(GLUvertex* u, GLUvertex* v, GLUvertex* w)
+GLdouble __gl_transEval(GLUvertex* u, GLUvertex* v, GLUvertex* w)
 {
    /* Given three vertices u,v,w such that TransLeq(u,v) && TransLeq(v,w),
     * evaluates the t-coord of the edge uw at the s-coord of the vertex v.
@@ -118,7 +118,7 @@ GLfloat __gl_transEval(GLUvertex* u, GLUvertex* v, GLUvertex* w)
     * let r be the negated result (this evaluates (uw)(v->t)), then
     * r is guaranteed to satisfy MIN(u->s,w->s) <= r <= MAX(u->s,w->s).
     */
-   GLfloat gapL, gapR;
+   GLdouble gapL, gapR;
 
    assert(TransLeq(u, v) && TransLeq(v, w));
 
@@ -141,13 +141,13 @@ GLfloat __gl_transEval(GLUvertex* u, GLUvertex* v, GLUvertex* w)
    return 0;
 }
 
-GLfloat __gl_transSign(GLUvertex* u, GLUvertex* v, GLUvertex* w)
+GLdouble __gl_transSign(GLUvertex* u, GLUvertex* v, GLUvertex* w)
 {
    /* Returns a number whose sign matches TransEval(u,v,w) but which
     * is cheaper to evaluate.  Returns > 0, == 0 , or < 0
     * as v is above, on, or below the edge uw.
     */
-   GLfloat gapL, gapR;
+   GLdouble gapL, gapR;
 
    assert(TransLeq(u, v) && TransLeq(v, w));
 
@@ -199,7 +199,7 @@ int __gl_vertCCW(GLUvertex* u, GLUvertex* v, GLUvertex* w)
 #include <stdlib.h>
 extern int RandomInterpolate;
 
-GLfloat Interpolate(GLfloat a, GLfloat x, GLfloat b, GLfloat y)
+GLdouble Interpolate(GLdouble a, GLdouble x, GLdouble b, GLdouble y)
 {
    printf("*********************%d\n",RandomInterpolate);
    if (RandomInterpolate)
@@ -224,7 +224,7 @@ void __gl_edgeIntersect(GLUvertex* o1, GLUvertex* d1,
  * bounding rectangles defined by each edge.
  */
 {
-   GLfloat z1, z2;
+   GLdouble z1, z2;
 
    /* This is certainly not the most efficient way to find the intersection
     * of two line segments, but it is very numerically stable.
